02-tool-use/requirements.txt
# Core dependencies
anthropic>=0.40.0
openai>=1.54.0
python-dotenv>=1.0.0

# Schema validation
pydantic>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Logging
loguru>=0.7.0
